requirements.txt
asgiref==3.5.2
beautiful-date==2.2.0
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
distlib==0.3.5
Django==4.0.6
django-extensions==3.2.0
django-widget-tweaks==1.4.12
docopt==0.6.2
filelock==3.7.1
gcsa==1.3.0
google-api-core==2.8.2
google-api-python-client==2.55.0
google-apis-oauth-django==1.0.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.56.4
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
MarkupSafe==2.1.1
oauthlib==3.2.0
pipreqs==0.4.11
platformdirs==2.5.2
protobuf==4.21.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
tzlocal==2.1
uritemplate==4.1.1
urllib3==1.26.11
virtualenv==20.15.1
Werkzeug==2.2.1
whitenoise==6.2.0
yarg==0.1.9
